refactor(core,logos_app): remove duplicated core_manager
