+++ core/src/core_manager/core_manager_interface.h
@@ -3,7 +3,7 @@
 
 #include <QString>
 #include <QStringList>
-#include "../interface.h"
+#include "../../interface.h"
 
 // Core manager specific methods interface
 class CoreManagerInterface {
