+++ logos_app/app/main.cpp
@@ -8,29 +8,39 @@
 #include <QStringList>
 #include <QDebug>
 #include <QMetaObject>
-#include "core_manager.h"
-#include "../../core/plugin_registry.h"
+
+// Replace CoreManager with direct C API functions
+extern "C" {
+    void logos_core_set_plugins_dir(const char* plugins_dir);
+    void logos_core_start();
+    void logos_core_cleanup();
+    char** logos_core_get_loaded_plugins();
+    int logos_core_load_plugin(const char* plugin_name);
+}
+
+// Helper function to convert C-style array to QStringList
+QStringList convertPluginsToStringList(char** plugins) {
+    QStringList result;
+    if (plugins) {
+        for (int i = 0; plugins[i] != nullptr; i++) {
+            result.append(plugins[i]);
+        }
+    }
+    return result;
+}
 
 int main(int argc, char *argv[])
 {
     // Create QApplication first
     QApplication app(argc, argv);
 
-
-    // TODO: this usage of CoreManager seems incorrect, as it's likely creating two instances, one here
-    // and another in the core_manager plugin itself, so things like setPluginsDirectory is not working properly
-
-    // Initialize CoreManager
-    CoreManager& core = CoreManager::instance();
-    core.initialize(argc, argv);
-
     // Set the plugins directory
     QString pluginsDir = QDir::cleanPath(QCoreApplication::applicationDirPath() + "/bin/plugins");
     std::cout << "Setting plugins directory to: " << pluginsDir.toStdString() << std::endl;
-    core.setPluginsDirectory(pluginsDir);
+    logos_core_set_plugins_dir(pluginsDir.toUtf8().constData());
 
     // Start the core
-    core.start();
+    logos_core_start();
     std::cout << "Logos Core started successfully!" << std::endl;
 
     // Load package_manager plugin by default
@@ -42,7 +52,9 @@ int main(int argc, char *argv[])
     }
 
     // Print loaded plugins initially
-    QStringList plugins = core.getLoadedPlugins();
+    char** loadedPlugins = logos_core_get_loaded_plugins();
+    QStringList plugins = convertPluginsToStringList(loadedPlugins);
+
     if (plugins.isEmpty()) {
         qInfo() << "No plugins loaded.";
     } else {
@@ -64,7 +76,7 @@ int main(int argc, char *argv[])
     int result = app.exec();
 
     // Cleanup
-    core.cleanup();
+    logos_core_cleanup();
 
     return result;
 }
